server/requirements.txt
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.9.0
boto3==1.37.18
botocore==1.37.18
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cws-helpers @ git+https://github.com/caseywschmid/cws-helpers.git@43f8e94357bd91dd9647bd8498f7da5dc58f548f
distro==1.9.0
fastapi==0.115.11
google-api-core==2.24.2
google-api-python-client==2.165.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.69.2
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jiter==0.9.0
jmespath==1.0.1
oauthlib==3.2.2
openai==1.68.2
proto-plus==1.26.1
protobuf==6.30.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.11.4
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.39
starlette==0.46.1
# Editable install with no version control (state-machine-demo==0.1.0)
-e /Users/caseyschmid/Programming/Discovitas/development/state_machine_demo/server
tqdm==4.67.1
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
yt-dlp==2025.3.21
